backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.0.1
websockets==12.0
celery==5.3.4
requests==2.31.0
email-validator==2.1.0 
